Least effort optimistic frame/packet alignment
This change factors all of the tracking logic for packets in one
direction into the session tracker.
This makes it possible for the tracker to recover from a frame parsing
error, and allows us to speculatively parse packets mid-stream.
In this pass, we assume that a good frame will start at some future
frame boundary.